clang tidy

+++ tasks/zavyalov_a_scalar_product/tests/performance/main.cpp
@@ -1,7 +1,5 @@
 #include <gtest/gtest.h>
 
-#include <iostream>
-
 #include "util/include/perf_test_util.hpp"
 #include "zavyalov_a_scalar_product/common/include/common.hpp"
 #include "zavyalov_a_scalar_product/mpi/include/ops_mpi.hpp"
@@ -10,16 +8,16 @@
 namespace zavyalov_a_scalar_product {
 
 class ZavyalovAScalarProductPerfTestProcesses : public ppc::util::BaseRunPerfTests<InType, OutType> {
-  const unsigned long long kCount_ = 30000000;
+  const uint64_t kCount_ = 30000000ULL;
   InType input_data_{};
 
   void SetUp() override {
     std::vector<double> leftVec(kCount_);
     std::vector<double> rightVec(kCount_);
 
-    for (unsigned long long i = 0; i < kCount_; i++) {
+    for (uint64_t i = 0; i < kCount_; i++) {
       leftVec[i] = i;
-      rightVec[i] = i * 2ull;
+      rightVec[i] = i * 2ULL;
     }
 
     input_data_ = std::make_tuple(leftVec, rightVec);
@@ -27,7 +25,7 @@ class ZavyalovAScalarProductPerfTestProcesses : public ppc::util::BaseRunPerfTes
 
   bool CheckTestOutputData(OutType &output_data) final {
     double res = 0.0;
-    for (unsigned long long i = 0; i < kCount_; i++) {
+    for (uint64_t i = 0; i < kCount_; i++) {
       res += std::get<0>(input_data_)[i] * std::get<1>(input_data_)[i];
     }
     double diff = fabs(res - output_data);
